move: internal/gui -> internal/ui
