+++ api/src/test/kotlin/online/aruka/custom_crafter/api/ui/AllCandidateUITest.kt
@@ -6,8 +6,8 @@ import io.github.sakaki_aruka.customcrafter.recipe.CoordinateComponent
 import io.github.sakaki_aruka.customcrafter.search.Search
 import io.github.sakaki_aruka.customcrafter.matter.CMatterImpl
 import io.github.sakaki_aruka.customcrafter.recipe.CRecipeImpl
-import io.github.sakaki_aruka.customcrafter.internal.gui.allcandidate.AllCandidateUI
-import io.github.sakaki_aruka.customcrafter.internal.gui.crafting.CraftUI
+import io.github.sakaki_aruka.customcrafter.internal.ui.allcandidate.AllCandidateUI
+import io.github.sakaki_aruka.customcrafter.internal.ui.crafting.CraftUI
 import io.github.sakaki_aruka.customcrafter.ui.AllCandidateUIDesigner
 import org.bukkit.Material
 import org.bukkit.entity.Player
